+++ extension/pathfind.h
@@ -5,6 +5,6 @@
 #include "entity.h"
 #include "state.h"
 
-Vec2f* path(Vec2f playerPosition, Entity** entities, GameState* gState);
+Vec2d* path(Vec2d playerPosition, Entity** entities, GameState* gState);
 
 #endif //PATHFIND_H
